Reduce duplication in server
